+++ angular/src/app/dataset/dataset-import/dataset-import.component.html
@@ -19,7 +19,7 @@
 </div>
 
 <div *ngIf="datasetList$ | async">
-  <ddap-dataset-list
+  <ddap-dataset-results
     [dataset$]=datasetList$
-  ></ddap-dataset-list>
+  ></ddap-dataset-results>
 </div>
