+++ angular/src/app/dataset/dataset.module.ts
@@ -5,15 +5,15 @@ import { MatFormFieldModule } from '@angular/material/form-field';
 
 import { SharedModule } from '../shared/shared.module';
 
-import { DatasetListComponent } from './dataset-list/dataset-list.component';
+import { DatasetImportComponent } from './dataset-import/dataset-import.component';
+import { DatasetResultsComponent } from './dataset-results/dataset-results.component';
 import { DatasetRoutingModule } from './dataset-routing.module';
-import { DatasetSearchComponent } from './dataset-search/dataset-search.component';
 import { DatasetViewAccessComponent } from './dataset-view-access/dataset-view-access.component';
 import { DatasetViewsComponent } from './dataset-views/dataset-views.component';
 
 
 @NgModule({
-  declarations: [DatasetSearchComponent, DatasetListComponent, DatasetViewsComponent, DatasetViewAccessComponent],
+  declarations: [DatasetImportComponent, DatasetResultsComponent, DatasetViewsComponent, DatasetViewAccessComponent],
   imports: [
     CommonModule,
     DatasetRoutingModule,
